Refactor models and storage structures
- Remove the old `models.rs` file and replace it with a new `mod.rs` that defines updated data structures for tasks and categories, including a priority enum.
- Update the `storage` module to utilize the new `StorageData`, `Task`, `Category`, and `Priority` structures, ensuring consistency across the application.
- Enhance data structures with additional fields for better tracking of task and category metadata.
